Roughly working refactor
