+++ src/containers/Elements/index.js
@@ -1,5 +1,4 @@
-export { default as NodePanel } from './NodePanel';
-export { default as NodeProviderPanels } from './NodeProviderPanels';
+export { default as NodePanels } from './NodePanels';
 export { default as PromptSwiper } from './PromptSwiper';
 export { default as Dialog } from './Dialog';
 export { default as Form } from './Form';
